refactor(ecstore): use local replication paths
